requirements.txt
# Flask and web framework
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
Werkzeug==2.3.7

# Core dependencies
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0

# Audio processing
openai-whisper>=20231117
pydub>=0.25.1

# Computer Vision and Image Processing
opencv-python>=4.8.0
scikit-image>=0.21.0
scipy>=1.11.0

# OCR
easyocr>=1.7.0

# Machine Learning (required for MaskFormer and other models)
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0

# Object Detection
ultralytics>=8.0.0

# MedSAM Dependencies
supabase-py>=2.0.0
aiohttp>=3.8.0

# HuggingFace Space Client (for remote segmentation API)
gradio_client>=0.10.0

# OpenAI API
openai>=1.0.0

# Note: Some dependencies may need to be optimized for Vercel's size limits
# Consider using lighter alternatives or external APIs for heavy models